+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/Kafka/MessagingTags.cs
@@ -1,0 +1,44 @@
+﻿// <copyright file="MessagingTags.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Kafka;
+
+internal static class MessagingTags
+{
+    internal static class Keys
+    {
+        public const string MessagingSystem = "messaging.system";
+        public const string MessagingOperation = "messaging.operation";
+        public const string DestinationName = "messaging.destination.name";
+        public const string ClientId = "messaging.client_id";
+
+        internal static class Kafka
+        {
+            public const string Partition = "messaging.kafka.destination.partition";
+            public const string MessageKey = "messaging.kafka.message.key";
+            public const string PartitionOffset = "messaging.kafka.message.offset";
+            public const string IsTombstone = "messaging.kafka.message.tombstone";
+            public const string ConsumerGroupId = "messaging.kafka.consumer.group";
+        }
+    }
+
+    internal static class Values
+    {
+        public const string KafkaMessagingSystemName = "kafka";
+        public const string PublishOperationName = "publish";
+        public const string ProcessOperationName = "process";
+    }
+}